Probable crash if input boundary length is over 70 chars
The length storage for the boundary is 70 bytes, which is the max
boundary size given by the RFC 1341. If the input boundary is over
70 chars, the strncpy() no copy the final '\0' and the next strlen
cause a crash.

This patch controls the length and force final '\0'. the init function
return 0 if the init fail with a too long boundary. however, for
compatibility, the library remains initialized and could be used
without crash.